+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.iyxan23.blocks.view">
+    package="com.openblocks.blocks.view">
 
     <application
         android:allowBackup="true"
@@ -10,7 +10,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Sketchwareblocksview">
         <activity
-            android:name=".MainActivity"
+            android:name="com.openblocks.blocks.view.example.MainActivity"
             android:label="@string/app_name"
             android:theme="@style/Theme.Sketchwareblocksview.NoActionBar">
             <intent-filter>
