+++ app/src/main/java/com/openblocks/blocks/view/example/MainActivity.java
@@ -1,8 +1,10 @@
-package com.iyxan23.blocks.view;
+package com.openblocks.blocks.view.example;
 
 import android.os.Bundle;
 
 import androidx.appcompat.app.AppCompatActivity;
+
+import com.openblocks.blocks.view.R;
 
 public class MainActivity extends AppCompatActivity {
 
@@ -10,6 +12,5 @@ public class MainActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-
     }
 }
