+++ app/src/main/res/layout/activity_main.xml
@@ -4,9 +4,9 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".example.MainActivity">
 
-    <com.iyxan23.blocks.view.SketchwareBlocksView
+    <com.openblocks.blocks.view.SketchwareBlocksView
         android:id="@+id/blocks_view"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
