+++ app/src/test/java/com/openblocks/blocks/view/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.iyxan23.blocks.view;
+package com.openblocks.blocks.view;
 
 import org.junit.Test;
