+++ lib/src/main/java/com/openblocks/blocks/view/SketchwareBlocksView.java
@@ -1,9 +1,8 @@
-package com.iyxan23.blocks.view;
+package com.openblocks.blocks.view;
 
 import android.content.Context;
 import android.content.res.TypedArray;
 import android.graphics.Canvas;
-import android.graphics.Color;
 import android.graphics.Paint;
 import android.graphics.Typeface;
 import android.util.AttributeSet;
