+++ lib/src/main/java/com/openblocks/blocks/view/SketchwareEvent.java
@@ -1,4 +1,4 @@
-package com.iyxan23.blocks.view;
+package com.openblocks.blocks.view;
 
 import android.graphics.Canvas;
 import android.graphics.Paint;
@@ -6,7 +6,6 @@ import android.graphics.Paint;
 import androidx.annotation.NonNull;
 
 import java.util.ArrayList;
-import java.util.Arrays;
 
 public class SketchwareEvent {
