+++ lib/src/main/java/com/openblocks/blocks/view/SketchwareField.java
@@ -1,9 +1,7 @@
-package com.iyxan23.blocks.view;
+package com.openblocks.blocks.view;
 
 import android.graphics.Canvas;
 import android.graphics.Paint;
-
-import java.util.ArrayList;
 
 public class SketchwareField {
