+++ lib/src/main/java/com/openblocks/blocks/view/SketchwareNestedBlock.java
@@ -1,4 +1,4 @@
-package com.iyxan23.blocks.view;
+package com.openblocks.blocks.view;
 
 import java.util.ArrayList;
